fix: don't use global connection object in `AddToDatabasePipeline`
Before this commit, the AddToDatabasePipeline function was using the
global connection object to insert the data into the database. This
caused error (`pymysql.err.InterfaceError: (0, '')`) which probably
caused because of using the same connection object in multiple threads.

This commit fixes this issue by creating a new connection object item
that enters the `AddToDatabasePipeline`.